refactor(shared): create the useLocalStorage file
