+++ src/shared/common/hook/useLocalStorage/index.ts
@@ -1,22 +1,1 @@
-import { useEffect, useState } from "react";
-import { UseLocalStorage } from "./types/UseLocalStorage.types";
-
-const useLocalStorage = <T>(key: string, initialState?: T): UseLocalStorage<T> => {
-	const [state, setState] = useState(() => {
-		const storage = localStorage.getItem(key);
-        
-		if (storage) {
-			return JSON.parse(storage);
-		} else {
-			return initialState;
-		}
-	});
-
-	useEffect(() => {
-		localStorage.setItem(key, JSON.stringify(state));
-	}, [key, state]);
-
-	return [state, setState];
-};
-
-export default useLocalStorage;
+export { default } from "./useLocalStorage";
